Internal rewrite to use kseqpp under the hood
FQFeeder now uses kseqpp (C++) instead of kseq (C)
as the underlying parser.  They are very similar in
raw speed (kseqpp is sometimes slightly faster), but
since we return read and read pair structures that
contain C++ strings rather than raw char pointers, using
kseqpp allows us to avoid a copy of the underlying
sequence (and quality) to a C++ string from a char*.

Overall, this change results in a cleaner and more simplified
design, and a slight peformance boost.